backend/requirements.txt
django>=5.2.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.7.0
python-dotenv>=1.0.0
requests>=2.32.0
numpy>=2.3.0
scikit-learn>=1.7.0
